+++ app/src/main/java/org/simple/clinic/ClinicApp.kt
@@ -4,6 +4,7 @@ import android.annotation.SuppressLint
 import android.app.Application
 import androidx.camera.camera2.Camera2Config
 import androidx.camera.core.CameraXConfig
+import androidx.fragment.app.FragmentManager
 import io.reactivex.exceptions.UndeliverableException
 import io.reactivex.plugins.RxJavaPlugins
 import org.simple.clinic.activity.CloseActivitiesWhenUserIsUnauthorized
@@ -43,12 +44,14 @@ abstract class ClinicApp : Application(), CameraXConfig.Provider {
 
   protected open val crashReporterSinks = emptyList<CrashReporter.Sink>()
 
-  @SuppressLint("RestrictedApi")
+  @SuppressLint("RestrictedApi", "UnsafeOptInUsageWarning")
   override fun onCreate() {
     super.onCreate()
 
     appComponent = buildDaggerGraph()
     appComponent.inject(this)
+
+    FragmentManager.enableNewStateManager(false)
 
     crashReporterSinks.forEach(CrashReporter::addSink)
     Timber.plant(CrashBreadcrumbsTimberTree())