+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 - Bump AppCompat to v1.3.0
 - Move online patient lookup behind a feature flag
 - Cache build dependencies and intermediates in CI workflows
+- Bump Fragment to v1.3.6
+- Disable new `FragmentStateManager`
 - [In Progress: 21 Jul 2021] Use `MobiusLoopViewModel` for sending view effects that are lifecycle aware
   - Replace `MobiusLoop.Controller` with `MobiusLoopViewModel` in base screens
   - Add new interface for handling received view effects